==> notebooks/old_notebooks/0_scraping_AmbisonicSoundLibrary.ipynb <==
# https://library.soundfield.com/browse/Ambience/Nature
import requests
from bs4 import BeautifulSoup
import re
import pandas as pd


title_df_list = []
for nPage in range(1,13):
    
    page = requests.get("https://library.soundfield.com/browse/Ambience/Nature?page="+str(nPage))
    soup = BeautifulSoup(page.content, 'html.parser')
    # print(soup.prettify())
    title_df = pd.DataFrame(soup.find_all('div', class_='track_title'), columns = ['track_title'])
    
    url = []
    for n in range(len(title_df)):
        text = str(list(soup.find_all('div', class_='waveform'))[n])
        m = re.search('data-src="(.+?)" id', text)
        try:
            found = m.group(1)
        except AttributeError:
            found = 'NA'

        url.append(found)
    title_df['url'] = url

    title_df_list.append(title_df)
    del title_df

title_df_all = pd.concat(title_df_list, ignore_index=True)
title_df_all

title_df_all.to_csv('../data/raw/AmbisonicSoundLibrary/ASL_nature.csv')

import urllib.request
for index, row in title_df_all.iterrows():
    urllib.request.urlretrieve(row['url'], "../data/raw/AmbisonicSoundLibrary/nature/"+row['track_title']+".mp3")
