==> notebooks/old_notebooks/4-1_YAMNet_lazypredict.ipynb <==
import os

from IPython import display
# import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import random as python_random

import tensorflow as tf
import tensorflow_hub as hub
from tensorflow.keras import regularizers

from lazypredict.Supervised import LazyClassifier

from sklearn.impute import SimpleImputer
from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import Pipeline


yamnet_model_handle = 'https://tfhub.dev/google/yamnet/1'
yamnet_model = hub.load(yamnet_model_handle) # if there's an "SavedModel file does not exist at:", delete that folder and rerun it


df_all = pd.read_csv('../train_val_test_split/train_val_test_GoogleAudioSet.csv', index_col=0)
df_all

# check the weight
df_all.groupby(['source'])['weight'].sum()

def extract_yamnet_embedding(filename):
    import pickle
    
    file = open(filename, 'rb')
    output = pickle.load(file)
#     output = pd.read_pickle(file)
    wav_raw = output['y']
    wav_bg = output['bg_y']
    wav_fg = output['fg_y']
    file.close()

# # no need to run padding as the time dimension will be collapsed after transforming to embeddings anyway    
#     if len(wav_raw) < 160000:
#         wav_raw = np.pad(wav_raw, (0,160000-len(wav_raw)),'mean')

#     if len(wav_bg) < 160000:
#         wav_bg = np.pad(wav_bg, (0,160000-len(wav_bg)),'mean')

#     if len(wav_fg) < 160000:
#         wav_fg = np.pad(wav_fg, (0,160000-len(wav_fg)),'mean')

    # Extract YAMNet embeddings for each frame
    scores, embedding_tensor_raw, spectrogram = yamnet_model(wav_raw)
    embedding_tensor_raw = tf.reduce_mean(embedding_tensor_raw, axis=0).numpy()

    scores, embedding_tensor_bg, spectrogram = yamnet_model(wav_bg)
    embedding_tensor_bg = tf.reduce_mean(embedding_tensor_bg, axis=0).numpy()

    scores, embedding_tensor_fg, spectrogram = yamnet_model(wav_fg)
    embedding_tensor_fg = tf.reduce_mean(embedding_tensor_fg, axis=0).numpy()
        
    
    return embedding_tensor_raw, embedding_tensor_bg, embedding_tensor_fg

from joblib import Parallel, delayed

embeddings_raw_list = []
embeddings_bg_list = []
embeddings_fg_list = []
import time
start_time = time.time()


for index, row in df_all.iterrows():
    embedding_tensor_raw, embedding_tensor_bg, embedding_tensor_fg = extract_yamnet_embedding(row['file'])
    
    embeddings_raw_list.append(embedding_tensor_raw)
    embeddings_bg_list.append(embedding_tensor_bg)
    embeddings_fg_list.append(embedding_tensor_fg)
    
embeddings_raw_matrix = np.stack(embeddings_raw_list, axis=0)
embeddings_bg_matrix = np.stack(embeddings_bg_list, axis=0)
embeddings_fg_matrix = np.stack(embeddings_fg_list, axis=0)
    
print('seconds: '+str(time.time()-start_time))


embeddings_matrix = np.concatenate((embeddings_raw_matrix, embeddings_bg_matrix, embeddings_fg_matrix), axis=1)

train_index = df_all['fold']<8
valid_index = df_all['fold']==8
test_index = df_all['fold']==9

# Raw signal

X_train = embeddings_raw_matrix[train_index,:]
y_train = df_all.loc[train_index,'category']
X_valid = embeddings_raw_matrix[valid_index,:]
y_valid = df_all.loc[valid_index,'category']

transformer = Pipeline(steps=[
       ('imputer', SimpleImputer(missing_values=np.nan, strategy='mean'))
      ,('scaler', StandardScaler())
])

transformer.fit(X_train)
X_train = transformer.transform(X_train)
X_valid = transformer.transform(X_valid)


clf = LazyClassifier(verbose=0, ignore_warnings=True, custom_metric=None)
models,predictions = clf.fit(X_train, X_valid, y_train, y_valid)
models

# Background signal

X_train = embeddings_bg_matrix[train_index,:]
y_train = df_all.loc[train_index,'category']
X_valid = embeddings_bg_matrix[valid_index,:]
y_valid = df_all.loc[valid_index,'category']

transformer = Pipeline(steps=[
       ('imputer', SimpleImputer(missing_values=np.nan, strategy='mean'))
      ,('scaler', StandardScaler())
])

transformer.fit(X_train)
X_train = transformer.transform(X_train)
X_valid = transformer.transform(X_valid)


clf = LazyClassifier(verbose=0, ignore_warnings=True, custom_metric=None)
models,predictions = clf.fit(X_train, X_valid, y_train, y_valid)
models

# Foreground signal

X_train = embeddings_fg_matrix[train_index,:]
y_train = df_all.loc[train_index,'category']
X_valid = embeddings_fg_matrix[valid_index,:]
y_valid = df_all.loc[valid_index,'category']

transformer = Pipeline(steps=[
       ('imputer', SimpleImputer(missing_values=np.nan, strategy='mean'))
      ,('scaler', StandardScaler())
])

transformer.fit(X_train)
X_train = transformer.transform(X_train)
X_valid = transformer.transform(X_valid)


clf = LazyClassifier(verbose=0, ignore_warnings=True, custom_metric=None)
models,predictions = clf.fit(X_train, X_valid, y_train, y_valid)
models

# All signal

X_train = embeddings_matrix[train_index,:]
y_train = df_all.loc[train_index,'category']
X_valid = embeddings_matrix[valid_index,:]
y_valid = df_all.loc[valid_index,'category']

transformer = Pipeline(steps=[
       ('imputer', SimpleImputer(missing_values=np.nan, strategy='mean'))
      ,('scaler', StandardScaler())
])

transformer.fit(X_train)
X_train = transformer.transform(X_train)
X_valid = transformer.transform(X_valid)


clf = LazyClassifier(verbose=0, ignore_warnings=True, custom_metric=None)
models,predictions = clf.fit(X_train, X_valid, y_train, y_valid)
models


  
    
    
    

# # Define a function to extract YAMNet embeddings for a single audio file
# def extract_yamnet_embedding(audio_path):
#     # Load the audio file
#     audio, sample_rate = tf.audio.decode_wav(tf.io.read_file(audio_path))
    
#     # Extract YAMNet embeddings for each frame
#     embedding_tensor = yamnet_model(audio)
#     embeddings = tf.reduce_mean(embedding_tensor, axis=0)
    
#     return embeddings

# # Define a function to parse a single metadata file
# def parse_metadata(metadata_path):
#     # Implement your metadata parsing code here
#     metadata = ...
    
#     return metadata

# # Define a function to combine YAMNet embeddings and metadata for a single audio file
# def combine_features(audio_path, metadata_path):
#     # Extract YAMNet embeddings
#     yamnet_embedding = extract_yamnet_embedding(audio_path)
    
#     # Parse metadata
#     metadata = parse_metadata(metadata_path)
    
#     # Concatenate the YAMNet embedding and metadata into a single feature vector
#     feature_vector = tf.concat([yamnet_embedding, metadata], axis=-1)
    
#     return feature_vector

# # Define a function to load and preprocess a single data sample
# def load_and_preprocess_data(audio_path, metadata_path):
#     # Combine YAMNet embeddings and metadata
#     feature_vector = combine_features(audio_path, metadata_path)
    
#     # Implement your data preprocessing code here
#     preprocessed_data = ...
    
#     return preprocessed_data

# # Load a list of audio and metadata file paths
# audio_paths = ["path/to/audio1.wav", "path/to/audio2.wav", ...]
# metadata_paths = ["path/to/metadata1.csv", "path/to/metadata2.csv", ...]

# # Create a TensorFlow dataset using the audio and metadata file paths
# dataset = tf.data.Dataset.from_tensor_slices((audio_paths, metadata_paths))

# # Use the map function to apply the load_and_preprocess_data function to each sample in the dataset
# dataset = dataset.map(load_and_preprocess_data)

# # Implement your training code here using the preprocessed dataset
# ...
