==> notebooks/old_notebooks/6_MPS_data.ipynb <==
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import tensorflow as tf
import random as python_random

meta_df = pd.read_csv('../train_val_test_split/train_val_test_GoogleAudioSet.csv', index_col=0)
meta_df

def load_mps_pkl(filename, mps_ind):
    import pickle
    import pandas as pd
    # open a file, where you stored the pickled data
    file = open(filename, 'rb')

    # dump information to that file
    output = pickle.load(file)
    
    mps = output[mps_ind]
    wt = output['wt']
    mps = mps[:,wt<=100].reshape(-1) # exclude the temporal modulation frequency >100 Hz

    # close the file
    file.close()
    return mps

mps_raw_list = []
mps_bg_list = []
mps_fg_list = []
for index, row in meta_df.iterrows():
    mps_raw_list.append(load_mps_pkl(row['file'], 'mps'))
    mps_bg_list.append(load_mps_pkl(row['file'], 'mps_bg'))
    mps_fg_list.append(load_mps_pkl(row['file'], 'mps_fg'))

from itertools import compress
train_raw = np.stack(compress(mps_raw_list, meta_df['fold']<8))
valid_raw = np.stack(compress(mps_raw_list, meta_df['fold']==8))

train_bg = np.stack(compress(mps_bg_list, meta_df['fold']<8))
valid_bg = np.stack(compress(mps_bg_list, meta_df['fold']==8))

train_fg = np.stack(compress(mps_fg_list, meta_df['fold']<8))
valid_fg = np.stack(compress(mps_fg_list, meta_df['fold']==8))

from lazypredict.Supervised import LazyClassifier
# from sklearn.impute import SimpleImputer
# from sklearn.preprocessing import StandardScaler
# from sklearn.pipeline import Pipeline
# from sklearn.decomposition import PCA

def get_transformer():
    from sklearn.preprocessing import StandardScaler
    from sklearn.pipeline import Pipeline
    from sklearn.decomposition import PCA
    transformer = Pipeline(steps=[
        ('PCA', PCA(n_components = 10, random_state=23)),
        ('scaler', StandardScaler())
    ])
    return transformer

# raw signal

transformer = get_transformer()
transformer.fit(train_raw)
train_raw_pca = transformer.transform(train_raw)
valid_raw_pca = transformer.transform(valid_raw)

y_train = meta_df.loc[meta_df['fold']<8, 'category']
y_valid = meta_df.loc[meta_df['fold']==8, 'category']


clf = LazyClassifier(verbose=0, ignore_warnings=True, custom_metric=None)
models,predictions = clf.fit(train_raw_pca, valid_raw_pca, y_train, y_valid)
models

# bg signal

transformer = get_transformer()
transformer.fit(train_bg)
train_bg_pca = transformer.transform(train_bg)
valid_bg_pca = transformer.transform(valid_bg)

y_train = meta_df.loc[meta_df['fold']<8, 'category']
y_valid = meta_df.loc[meta_df['fold']==8, 'category']


clf = LazyClassifier(verbose=0, ignore_warnings=True, custom_metric=None)
models,predictions = clf.fit(train_bg_pca, valid_bg_pca, y_train, y_valid)
models

# fg signal

transformer = get_transformer()
transformer.fit(train_fg)
train_fg_pca = transformer.transform(train_fg)
valid_fg_pca = transformer.transform(valid_fg)

y_train = meta_df.loc[meta_df['fold']<8, 'category']
y_valid = meta_df.loc[meta_df['fold']==8, 'category']


clf = LazyClassifier(verbose=0, ignore_warnings=True, custom_metric=None)
models,predictions = clf.fit(train_fg_pca, valid_fg_pca, y_train, y_valid)
models
